Arbitrary node distribution and mesh input (#1)
* Added capability to generate and mesh a cylindrical duct from parameters in the config file

* Simplified simpleBox function

Made similar to how the cylinder is generated. A meshed surface is extruded rather than creating the whole box and then meshing. Complicated for loop logics also removed since the box shape is constant (only dimensions change)

* Added wrapper function for maketestdomain

* Moved the Input class into it's own file

* Added functionality to take a mesh file as input

Coordinates of inlet nodes are taken from the nodes labelled as the inlet within the mesh file. These nodes are then used to discretise the flow field in core.py.

* Removed transformation from swirlgenerator axis to boundary condition axis since the axis systems are now the same. 

* Edited module interactions to match new expected workflow

If requested, a mesh file should be generated first and the node coordinates of the FlowField object are extracted from this. Also compatible for when a pre-made mesh is being used - by specifying it in the config file and not using the -makemesh option at run

* No longer creates a grid of nodes, uses the nodes in the input mesh

More representive of the actual use case of the tool. A pre-made mesh will be used as input to the tool and a boundary condition will be created on its inlet. However can still create a simple test mesh if one is not available

* Boundary curve function now generalised for most basic inlet shapes using alpha shape algorithm

A small value of alpha greater than 0 seems to correctly get the nodes at the boundary of most simple shapes. A value of 0 gives the convex hull which ignores the nodes at the sides of rectangles (outputting only the corner points).
The alpha parameter is hard coded as 0.1, for more complex inlet shapes this may need to be editable in the future. A call to alphashape.optimizealpha would be needed, but this takes a long time.

* Boundary function simplified and only activated for circular inlets

* Removed reliance of the core.py functions on the mesh parameters defined in the config file

* Removed non-functional copy function in FlowField

* Fixed check boundaries not actually calculating normal velocity correctly

* Converted post.py to object oriented and fixed plotting capability for new node storage

Moved plotting function into a class. Plotting functions can now handle irregularly spaced data points by interpolating them into a regular grid.

* Updated config file format and reading capability to match representative workflow

To reflect the fact that swirl generator can now be run with only a mesh input file and vortices description